+++ Code/backend/additionalTests/test1.js
@@ -1,4 +1,4 @@
-const request = require("supertest")("http://localhost:5000/api/v1");
+/*const request = require("supertest")("http://localhost:8000/api/v1");
 const expect = require("chai").expect;
 
 describe("GET /recipes", function () {
@@ -66,3 +66,4 @@ describe("GET /recipes", function () {
     expect(response.body.filters.CleanedIngredients).to.eql("pear");
   });
 });
+*/
